requirements.txt
anthropic
python-dotenv
requests
rich
beautifulsoup4
html2text
